# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: openai
    description: OpenAI AI chat, image, and audio operations
  - name: codegen
    description: Code generation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/check-username:
    post:
      operationId: checkUsername
      tags: [auth]
      summary: Check if a username is available
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckUsernameBody"
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckUsernameResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
  /openai/conversations:
    get:
      operationId: listOpenaiConversations
      tags: [openai]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiConversation"
    post:
      operationId: createOpenaiConversation
      tags: [openai]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpenaiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversation"
  /openai/conversations/{id}:
    get:
      operationId: getOpenaiConversation
      tags: [openai]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
    delete:
      operationId: deleteOpenaiConversation
      tags: [openai]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenaiError"
  /openai/conversations/{id}/messages:
    get:
      operationId: listOpenaiMessages
      tags: [openai]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenaiMessage"
    post:
      operationId: sendOpenaiMessage
      tags: [openai]
      summary: Send a text message and receive a streaming text response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOpenaiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant text chunks
          content:
            text/event-stream: {}
  /codegen/generate:
    post:
      operationId: generateCode
      tags: [codegen]
      summary: Generate code from a user description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCodeBody"
      responses:
        "200":
          description: SSE stream of generated code chunks
          content:
            text/event-stream: {}
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthError"
        "429":
          description: Daily limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CheckUsernameBody:
      type: object
      properties:
        username:
          type: string
      required:
        - username
    CheckUsernameResponse:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserInfo"
      required:
        - token
        - user
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        questionsToday:
          type: integer
        questionsLimit:
          type: integer
      required:
        - id
        - username
        - role
        - questionsToday
        - questionsLimit
    AuthError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    LimitError:
      type: object
      properties:
        error:
          type: string
        questionsToday:
          type: integer
        questionsLimit:
          type: integer
      required:
        - error
        - questionsToday
        - questionsLimit
    OpenaiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    OpenaiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateOpenaiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendOpenaiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    OpenaiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenaiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
    OpenaiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    GenerateCodeBody:
      type: object
      properties:
        prompt:
          type: string
        language:
          type: string
      required:
        - prompt