# lib/api-zod/src/generated/types/userInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserInfo {
  id: number;
  username: string;
  role: string;
  questionsToday: number;
  questionsLimit: number;
}